pyproject.toml
[tool.poetry]
name = "flask-test-study"
version = "0.1.0"
description = ""
authors = ["Léo Graciano <leonardoferreiragraciano@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_test_study"}]

[tool.poetry.dependencies]
python = "^3.8.0"
flask = "^2.2.2"
dynaconf = "^3.1.11"
flask-sqlalchemy = "^3.0.2"
flask-migrate = "^4.0.0"
flask-wtf = "^1.0.1"
flask-login = "^0.6.2"


[tool.poetry.group.dev.dependencies]
ward = "^0.67.0b0"
splinter = {version = "^0.18.1", extras = ["flask"]}
urllib3 = "^1.26.13"
autopep8 = "^2.0.0"
python-dotenv = "^0.21.0"
faker = "^15.3.4"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
